requirements-dev.txt
-r requirements.txt

django-stubs==1.12.0
django-stubs-ext==0.8.0
flake8==6.0.0
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
tomli==2.0.1
typing-extensions==4.5.0
types-PyYAML==6.0.12.9
types-pytz==2023.3.0.0
types-requests==2.28.11.17
types-urllib3==1.26.25.11
